支持对rtp流进行录制

+++ utils/stack.js
@@ -4,6 +4,7 @@
 const os = require('os')
 const fs = require('fs')
 const path = require('path')
+const tmp = require('tmp')
 const { nanoid } = require('nanoid')
 const ffmpeg = require('fluent-ffmpeg')
 
@@ -11,6 +12,9 @@ const RunningCommands = new Map()
 
 // 当前正在播放文件的软链接
 const PLAYING_SYMLINK = Symbol('playing_symlink')
+
+// 接收RTP进行录制时的临时SDP文件
+const PLAYING_SDPFILE = Symbol('playing_sdpfile')
 
 let playingDir = fs.mkdtempSync(path.join(os.tmpdir(), 'tms-koa-ffmpeg-'))
 
@@ -38,9 +42,16 @@ class CommandStack {
    */
   static removeCommand(uuid) {
     const cmd = RunningCommands.get(uuid)
-    if (cmd[PLAYING_SYMLINK]) {
-      if (fs.existsSync(cmd[PLAYING_SYMLINK]))
-        fs.unlinkSync(cmd[PLAYING_SYMLINK])
+    if (cmd) {
+      if (cmd[PLAYING_SYMLINK]) {
+        if (fs.existsSync(cmd[PLAYING_SYMLINK]))
+          fs.unlinkSync(cmd[PLAYING_SYMLINK])
+      }
+      if (cmd[PLAYING_SDPFILE] && cmd[PLAYING_SDPFILE].name) {
+        if (fs.existsSync(cmd[PLAYING_SDPFILE].name)) {
+          fs.unlinkSync(cmd[PLAYING_SDPFILE].name)
+        }
+      }
     }
 
     return RunningCommands.delete(uuid)
@@ -83,6 +94,26 @@ class CommandStack {
    */
   static getPlayingSymlink(cmd) {
     return cmd[PLAYING_SYMLINK]
+  }
+  /**
+   * 创建临时sdp文件
+   *
+   * @param {*} cmd
+   * @param {*} fullpath
+   */
+  static createPlayingSdpfile(cmd) {
+    const tmpSdp = tmp.fileSync()
+
+    cmd[PLAYING_SDPFILE] = tmpSdp
+
+    return tmpSdp.name
+  }
+  /**
+   *
+   * @param {*} cmd
+   */
+  static getPlayingSdpfile(cmd) {
+    return cmd[PLAYING_SDPFILE] ? cmd[PLAYING_SDPFILE].name : null
   }
 }
